created different required components for events page
