+++ src/pages/Events.jsx
@@ -1,19 +1,11 @@
-import React, { useState, useRef } from 'react';
+import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { 
-  faCalendarPlus, 
-  faEllipsisVertical, 
-  faMagnifyingGlass, 
-  faFilter, 
-  faLocationDot, 
-  faCalendarDays, 
-  faTimes,
-  faSave,
-  faCheck,
-  faSort,
-  faEye
-} from '@fortawesome/free-solid-svg-icons';
+import EventsHeader from '../components/EventsHeader';
+import EventCard from '../components/EventCard';
+import CreateEventModal from '../components/CreateEventModal';
+import FilterPanel from '../components/FilterPanel';
+import NoEventsMessage from '../components/NoEventsMessage';
 import { ToastContainer, toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 import '../stylesheets/events.css';
@@ -23,8 +15,7 @@ function EventsPage() {
   const [searchTerm, setSearchTerm] = useState('');
   const [showCreateForm, setShowCreateForm] = useState(false);
   const [showFilters, setShowFilters] = useState(false);
-  const modalRef = useRef(null);
-  
+
   // Form data for creating a new event
   const [newEventData, setNewEventData] = useState({
     name: '',
@@ -256,13 +247,6 @@ function EventsPage() {
     }
   };
 
-  // Close modal when clicking outside
-  const handleClickOutside = (e) => {
-    if (modalRef.current && !modalRef.current.contains(e.target)) {
-      handleCloseForm();
-    }
-  };
-
   return (
     <div className="events-page-background">
       <ToastContainer
@@ -277,280 +261,52 @@ function EventsPage() {
         pauseOnHover
         theme="colored"
       />
-      
       <div className="events-container">
-        <div className="events-header">
-          <h1 className="events-title">
-            <FontAwesomeIcon icon={faCalendarDays} className="title-icon" /> Events
-          </h1>
-          <button 
-            className="create-event-button" 
-            onClick={handleCreateEvent}
-          >
-            <FontAwesomeIcon icon={faCalendarPlus} className="button-icon" />
-            Create Event
-          </button>
-        </div>
-
-        <div className="events-filters">
-          <div className="search-container">
-            <FontAwesomeIcon icon={faMagnifyingGlass} className="search-icon" />
-            <input
-              type="text"
-              placeholder="Search events..."
-              value={searchTerm}
-              onChange={handleSearch}
-              className="search-input"
-            />
-          </div>
-          <button className="filter-button" onClick={handleToggleFilters}>
-            <FontAwesomeIcon icon={faFilter} />
-            <span>Filter & Sort</span>
-          </button>
-        </div>
+        {/* Header with Title, Create Button, and Search/Filter Controls */}
+        <EventsHeader 
+          onCreateEvent={handleCreateEvent}
+          onToggleFilters={handleToggleFilters}
+          searchTerm={searchTerm}
+          onSearch={handleSearch}
+        />
 
         {/* Filter Panel */}
         {showFilters && (
-          <div className="filter-panel">
-            <div className="filter-panel-header">
-              <h3>Filter & Sort Events</h3>
-              <button className="close-button" onClick={handleToggleFilters}>
-                <FontAwesomeIcon icon={faTimes} />
-              </button>
-            </div>
-            
-            <div className="filter-panel-content">
-              <div className="filter-group">
-                <label className="filter-label">Status</label>
-                <select 
-                  name="status" 
-                  className="filter-select"
-                  value={filters.status}
-                  onChange={handleFilterChange}
-                >
-                  <option value="all">All Statuses</option>
-                  <option value="planning">Planning</option>
-                  <option value="upcoming">Upcoming</option>
-                  <option value="completed">Completed</option>
-                </select>
-              </div>
-              
-              <div className="filter-group">
-                <label className="filter-label">Date Range</label>
-                <select 
-                  name="dateRange" 
-                  className="filter-select"
-                  value={filters.dateRange}
-                  onChange={handleFilterChange}
-                >
-                  <option value="all">All Dates</option>
-                  <option value="upcoming30">Next 30 Days</option>
-                  <option value="upcoming90">Next 90 Days</option>
-                  <option value="past">Past Events</option>
-                </select>
-              </div>
-              
-              <div className="filter-group">
-                <label className="filter-label">
-                  <FontAwesomeIcon icon={faSort} /> Sort By
-                </label>
-                <select 
-                  name="sortBy" 
-                  className="filter-select"
-                  value={filters.sortBy}
-                  onChange={handleFilterChange}
-                >
-                  <option value="date">Date (Oldest First)</option>
-                  <option value="name">Name (A-Z)</option>
-                  <option value="status">Status</option>
-                </select>
-              </div>
-            </div>
-            
-            <div className="filter-panel-footer">
-              <button 
-                className="apply-filters-button"
-                onClick={handleToggleFilters}
-              >
-                <FontAwesomeIcon icon={faCheck} /> Apply Filters
-              </button>
-            </div>
-          </div>
+          <FilterPanel
+            filters={filters}
+            onClose={handleToggleFilters}
+            onChange={handleFilterChange}
+            onApply={handleToggleFilters}
+          />
         )}
 
-        {/* Events Grid */}
+        {/* Events List */}
         <div className="events-list">
           {sortedEvents.length > 0 ? (
-            sortedEvents.map(event => (
-              <div className="event-card" key={event.id}>
-                <div className="event-card-content">
-                  <div className="event-header">
-                    <h3 className="event-name">{event.name}</h3>
-                    <span className={`event-status ${getStatusBadgeClass(event.status)}`}>
-                      {event.status.charAt(0).toUpperCase() + event.status.slice(1)}
-                    </span>
-                  </div>
-                  
-                  <div className="event-dates">
-                    <FontAwesomeIcon icon={faCalendarDays} className="event-icon" />
-                    {event.startDate === event.endDate ? (
-                      <p>{formatDate(event.startDate)}</p>
-                    ) : (
-                      <p>{formatDate(event.startDate)} - {formatDate(event.endDate)}</p>
-                    )}
-                  </div>
-                  
-                  <p className="event-location">
-                    <FontAwesomeIcon icon={faLocationDot} className="event-icon" />
-                    {event.location}
-                  </p>
-                  <p className="event-description">{event.description}</p>
-                </div>
-                
-                <div className="event-actions">
-                  <button 
-                    className="manage-event-button"
-                    onClick={() => handleManageEvent(event.id)}
-                  >
-                    Manage
-                  </button>
-                  <div className="event-menu-dropdown">
-                    <button className="event-menu-button">
-                      <FontAwesomeIcon icon={faEllipsisVertical} />
-                    </button>
-                    <div className="event-menu-content">
-                      <button onClick={() => handleViewDetails(event)}>
-                        <FontAwesomeIcon icon={faEye} /> View Details
-                      </button>
-                    </div>
-                  </div>
-                </div>
-              </div>
+            sortedEvents.map((event) => (
+              <EventCard
+                key={event.id}
+                event={event}
+                onManage={handleManageEvent}
+                onViewDetails={handleViewDetails}
+                getStatusBadgeClass={getStatusBadgeClass}
+                formatDate={formatDate}
+              />
             ))
           ) : (
-            <div className="no-events-message">
-              <FontAwesomeIcon icon={faCalendarDays} size="2x" style={{ marginBottom: '16px', opacity: 0.5 }} />
-              <p>No events found. Try adjusting your search criteria or create a new event.</p>
-              <button 
-                className="create-event-button-alt" 
-                onClick={handleCreateEvent}
-              >
-                <FontAwesomeIcon icon={faCalendarPlus} className="button-icon" />
-                Create Event
-              </button>
-            </div>
+            <NoEventsMessage onCreateEvent={handleCreateEvent} />
           )}
         </div>
       </div>
 
       {/* Create Event Modal */}
       {showCreateForm && (
-        <div className="modal-overlay" onClick={handleClickOutside}>
-          <div className="create-event-modal" ref={modalRef}>
-            <div className="modal-header">
-              <h2>Create New Event</h2>
-              <button className="close-modal-button" onClick={handleCloseForm}>
-                <FontAwesomeIcon icon={faTimes} />
-              </button>
-            </div>
-            
-            <form onSubmit={handleSubmitEvent} className="create-event-form">
-              <div className="form-row">
-                <div className="form-group">
-                  <label htmlFor="name">Event Name <span className="required">*</span></label>
-                  <input
-                    type="text"
-                    id="name"
-                    name="name"
-                    value={newEventData.name}
-                    onChange={handleFormChange}
-                    placeholder="Enter event name"
-                    required
-                  />
-                </div>
-              </div>
-              
-              <div className="form-row two-columns">
-                <div className="form-group">
-                  <label htmlFor="startDate">Start Date <span className="required">*</span></label>
-                  <input
-                    type="date"
-                    id="startDate"
-                    name="startDate"
-                    value={newEventData.startDate}
-                    onChange={handleFormChange}
-                    required
-                  />
-                </div>
-                <div className="form-group">
-                  <label htmlFor="endDate">End Date</label>
-                  <input
-                    type="date"
-                    id="endDate"
-                    name="endDate"
-                    value={newEventData.endDate}
-                    onChange={handleFormChange}
-                    min={newEventData.startDate} // End date must be after start date
-                  />
-                </div>
-              </div>
-              
-              <div className="form-row">
-                <div className="form-group">
-                  <label htmlFor="location">Location <span className="required">*</span></label>
-                  <input
-                    type="text"
-                    id="location"
-                    name="location"
-                    value={newEventData.location}
-                    onChange={handleFormChange}
-                    placeholder="Enter event location"
-                    required
-                  />
-                </div>
-              </div>
-              
-              <div className="form-row">
-                <div className="form-group">
-                  <label htmlFor="status">Status</label>
-                  <select
-                    id="status"
-                    name="status"
-                    value={newEventData.status}
-                    onChange={handleFormChange}
-                  >
-                    <option value="planning">Planning</option>
-                    <option value="upcoming">Upcoming</option>
-                    <option value="completed">Completed</option>
-                  </select>
-                </div>
-              </div>
-              
-              <div className="form-row">
-                <div className="form-group">
-                  <label htmlFor="description">Description</label>
-                  <textarea
-                    id="description"
-                    name="description"
-                    value={newEventData.description}
-                    onChange={handleFormChange}
-                    placeholder="Enter event description"
-                    rows="4"
-                  />
-                </div>
-              </div>
-              
-              <div className="form-actions">
-                <button type="button" className="cancel-button" onClick={handleCloseForm}>
-                  Cancel
-                </button>
-                <button type="submit" className="save-button">
-                  <FontAwesomeIcon icon={faSave} /> Create Event
-                </button>
-              </div>
-            </form>
-          </div>
-        </div>
+        <CreateEventModal
+          newEventData={newEventData}
+          onClose={handleCloseForm}
+          onSubmit={handleSubmitEvent}
+          onChange={handleFormChange}
+        />
       )}
     </div>
   );
